RPIcomm1.c for the ESP32 side of the UDOO Key
This program is for the ESP32 side of the UDOO Key.  This program sends a signal from IO22 of the ESP32 to GPIO 0 of the RP2040.  There is a companion program of the same name for the RP2040 side that reads the information the RP2040 receives on GPIO 0 and mirrors the information to the RP2040 LED.