RPIcomm1.c for UDOO Key, ESP32 side
RPIcomm1.c is for the ESP32 side of the UDOO Key.  It works with RPIcomm1B.c for the RP2040 side.  RPIcomm1.c sends a signal from IO22 of the ESP32 to GPIO 0 of the RP2040.  RPIcomm1B.c on the RP2040 reads the signal sent to GPIO 0 and mirrors the result to the RP2040 LED.